requirements.txt
openai==0.10.3
python-dotenv==0.19.0